+++ app/src/main/java/com/feitianzhu/huangliwo/im/ConversationListFragment.java
@@ -69,7 +69,7 @@ public class ConversationListFragment extends EaseConversationListFragment {
 
    /* @Override
     public void onCreateContextMenu(ContextMenu menu, View v, ContextMenuInfo menuInfo) {
-        getActivity().getMenuInflater().inflate(R.menu.em_delete_message, menu); 
+        getActivity().getMenuInflater().inflate(R.menu.em_delete_message, menu);
     }*/
 
   /*  @Override
