+++ easeui/src/main/res/layout/ease_widget_title_bar.xml
@@ -15,8 +15,8 @@
 
         <ImageView
             android:id="@+id/left_image"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
+            android:layout_width="20dp"
+            android:layout_height="20dp"
             android:layout_centerInParent="true"
             android:scaleType="centerInside" />
     </RelativeLayout>
